+++ src/media/midi/mod.rs
@@ -5,18 +5,15 @@
 // safety
 // #![cfg_attr(feature = "safe_midi", forbid(unsafe_code))]
 
-mod error;
-
 crate::items! { // structural access: _mods, _all, _always
     #[allow(unused)]
     pub use _mods::*;
     #[allow(unused)] #[doc(hidden, no_inline)]
     pub use _always::*;
 
-    mod _mods {
-        pub use super::error::*;
+    mod _mods { #![allow(unused)]
     }
-    pub(super) mod _all {
+    pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
         pub use super::_mods::*;
     }